+++ backend/molgenis-emx2-tasks/src/main/java/org/molgenis/emx2/tasks/TaskService.java
@@ -1,5 +1,6 @@
 package org.molgenis.emx2.tasks;
 
+import java.net.URL;
 import java.util.Collection;
 import java.util.List;
 import java.util.Set;
@@ -11,7 +12,7 @@ public interface TaskService {
 
   String submitTaskFromName(String name, String parameters);
 
-  String submitTaskFromName(String name, String parameters, String url);
+  String submitTaskFromName(String name, String parameters, URL host);
 
   Set<String> getJobIds();
