+++ backend/molgenis-emx2-webapi/src/main/java/org/molgenis/emx2/web/TaskApi.java
@@ -3,7 +3,7 @@ package org.molgenis.emx2.web;
 import static org.molgenis.emx2.Constants.SYSTEM_SCHEMA;
 import static org.molgenis.emx2.FilterBean.f;
 import static org.molgenis.emx2.SelectColumn.s;
-import static org.molgenis.emx2.rdf.RDFUtils.extractHost;
+import static org.molgenis.emx2.utils.URIUtils.extractHost;
 import static org.molgenis.emx2.web.FileApi.addFileColumnToResponse;
 import static org.molgenis.emx2.web.MolgenisWebservice.getSchema;
 import static org.molgenis.emx2.web.MolgenisWebservice.sessionManager;
@@ -13,6 +13,8 @@ import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import java.io.IOException;
 import java.io.UnsupportedEncodingException;
+import java.net.MalformedURLException;
+import java.net.URL;
 import java.net.URLDecoder;
 import java.nio.charset.StandardCharsets;
 import org.molgenis.emx2.*;
@@ -78,7 +80,8 @@ public class TaskApi {
     return new ObjectMapper().writeValueAsString(taskSchedulerService.scheduledTaskNames());
   }
 
-  private static String postScript(Request request, Response response) {
+  private static String postScript(Request request, Response response)
+      throws MalformedURLException {
     if (request.params("schema") == null || getSchema(request) != null) {
       MolgenisSession session = sessionManager.getSession(request);
       String user = session.getSessionUser();
@@ -88,8 +91,8 @@ public class TaskApi {
       String name = URLDecoder.decode(request.params("name"), StandardCharsets.UTF_8);
       String parameters = request.body();
 
-      String url = extractHost(request.url());
-      String id = taskService.submitTaskFromName(name, parameters, url);
+      URL host = new URL(extractHost(request.url()));
+      String id = taskService.submitTaskFromName(name, parameters, host);
       return new TaskReference(id).toString();
     }
     throw new MolgenisException("Schema doesn't exist or permission denied");
